Migrate to Room DB, improve resource management, and enhance API handling
