+++ app/src/main/res/values/openrouter_api_key.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="openrouter_api_key">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</string>
+    <string name="openrouter_api_key">YOUR_API_KEY_HERE</string>
 </resources>
